Major overhaul
Entirely new implementation, added a broadcast receiver to handle
picture intents so that the images could be saved immediately after the
picture is taken rather than saving a list of pictures and saving later
on. Added the ability to take videos. Power drain may occur in this
version but more testing need to be done to assess if/ how big of a
problem this is. This implementation does not yet work on all devices,
it does work on my HTC One S phone perfectly well but broader support
still needs to be achieved.